Initial Commit
Grater is a simple automatic, random mouse mover written in Rust. The sole purpose of this application is to programmatically simulate random mouse cursor movement to prevent a computer from going idle/to sleep.

This first commit establishes the fundamental logic for this singular task, as well as graceful and human-readible error handling and program runtime information to the console.